Comments, Fourgrams, Fivegrams, and comparative analysis
Added comments to make code more readable, implemented 4-grams and 5-grams for a better understanding of the impact of  overall accuracy v. case specificity and added the meat and potatoes, cross-corpus model testing with associated analyses regarding the frequency of specified n-grams in the corpus, and predictive accuracy (accuracy being determined both across the entirety of the target corpus and solely in instances were the target word is a possibility considered by the model).